avoid generating duplicate enum classess
Even though enums are placed in the file associated with a particular
class, they are generated in the global model namespace. So, we can't have
two enums with the same name.

This is a problem when we logically use the same enum, for the same
logical property in two different schemas, as they really should be named
the same.

To work around this issue, this change will keep track of enums that we've
already generated, and avoid generating the same enum again. If the enums
are not actually the same, just named the same, an error is thrown
instead.